#308 quit model injection
